requirements.txt
#$ conda install -c plotly plotly-orca ################ Have to install plotly-orca for saving output images###############Here is colab code [ ###Installing plotly-orca library 
										!pip install plotly>=4.7.1
										!wget https://github.com/plotly/orca/releases/download/v1.2.1/orca-1.2.1-x86_64.AppImage -O /usr/local/bin/orca
										!chmod +x /usr/local/bin/orca
										!apt-get install xvfb libgtk2.0-0 libgconf-2-4 ]
kneed
sentence-transformers
transformers
bert-extractive-summarizer
pandas
numpy
os
json
torch
matplotlib
sklearn
plotly
nltk
gensim
spacy
re
